Frame emitter managed by session as it should be...configurator now manages rendered rate
